js restruct
